# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IAM API REST in Golang with Gin Framework",
        "title": "eHealthcare Routing Service",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "iam-service",
    "basePath": "/v1",
    "paths": {
        "/actions": {
            "get": {
                "description": "Get available action from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Get Available Actions",
                "operationId": "get-action",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Action"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http_err.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Post new action to DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Post new action",
                "operationId": "post-action",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Action"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http_err.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http_err.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "models.Action": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                }
            }
        }
    }
}